Initial Mojang mappings remap
This branch of Loom has been completely remapped to use the Mojang mappings instead of the Fabric Yarn mappings we were using before.

While this branch does run and compile, there is still a lot of API cleanup that needs to be done.

+++ api/src/main/java/org/loomdev/api/entity/Explosive.java
@@ -17,6 +17,5 @@ public interface Explosive {
      *
      * @param power The new power of the explosion.
      */
-    void setExplosionPower(float power);
-
+    void setExplosionPower(int power);
 }
